grobal-navigationファイルの整理
