+++ src/components/navigation/global-navigation/pc-navigation/style.module.css
@@ -58,62 +58,3 @@
     margin-top: 24px;
   }
 }
-
-.navigationSp {
-  display: none;
-
-  .logo {
-    height: 48px;
-    width: 100%;
-    object-fit: contain;
-  }
-
-  .navigationList {
-    margin-top: 16px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 16px;
-
-    .navigationItem {
-      display: flex;
-      align-items: center;
-      background-color: transparent;
-      border-radius: 100px;
-      padding: 6px 16px 10px;
-      text-decoration: none;
-
-      &:hover {
-        background-color: var(--color-primary-dark);
-        opacity: 0.7;
-        transition: opacity 0.3s ease;
-      }
-    }
-
-    .current {
-      background-color: var(--color-white);
-
-      &:hover {
-        background-color: var(--color-primary-light);
-        opacity: 1;
-        transition: opacity 0.3s ease;
-      }
-    }
-  }
-
-  .social {
-    margin-top: 32px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 48px;
-  }
-
-  .copyright {
-    margin-top: 24px;
-  }
-
-  @media (max-width: 900px) {
-    display: block;
-  }
-}
